finalizados alguns tratamentos de exception

+++ src/main/java/com/example/crudteste/service/InstrumentoService.java
@@ -3,6 +3,9 @@ package com.example.crudteste.service;
 import com.example.crudteste.entity.InstrumentoEntity;
 import com.example.crudteste.repository.InstrumentoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Repository;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -14,7 +17,7 @@ public class InstrumentoService {
     @Autowired
     InstrumentoRepository instRepository;
 
-    public List<InstrumentoEntity> listarInstrumentos(){
+    public List<InstrumentoEntity> listarInstrumentos() {
         //        for (int i = 0; i < instRepository.findAll().size(); i++) {
 //            InstrumentoEntity instrumento = instrumentoEntityList.get(i);
 //            instRepository.save(instrumento);
@@ -24,18 +27,34 @@ public class InstrumentoService {
         //List<InstrumentoEntity> instrumentoRepository = instRepository.findAll();
         //return instrumentoRepository;
     }
-    public InstrumentoEntity adicionarInst(InstrumentoEntity instrumento){
-        return instRepository.save(instrumento);
+
+    public ResponseEntity<InstrumentoEntity> adicionarInst(InstrumentoEntity instrumento) {
+        if (null == instrumento.getDescricao()) {
+            return new ResponseEntity<>(HttpStatus.valueOf(400));
+        }
+        return new ResponseEntity<>(instRepository.save(instrumento), HttpStatus.CREATED);
     }
 
-    public void deleteInst(Long id) {
-        instRepository.deleteById(id);
+    public ResponseEntity<String> deleteInst(Long id) {
+        if (instRepository.existsById(id)) {
+            instRepository.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>("O ID ".concat(id.toString()).concat(" não existe! "),HttpStatus.NOT_FOUND);
+
     }
 
-    public InstrumentoEntity updateInst(Long instId, String editMarca){
-        InstrumentoEntity entity = instRepository.getById(instId);
-        entity.setMarca(editMarca);
-        InstrumentoEntity save = instRepository.save(entity);
-        return save;
+    public ResponseEntity<InstrumentoEntity> updateInst(Long instId, String editMarca) {
+        /*  Condicional que verifica se existe alguma entidade no banco de dados com esse Id    */
+        if(instRepository.existsById(instId)){
+            if(!editMarca.isBlank() || !editMarca.isEmpty()){
+
+                InstrumentoEntity entity = instRepository.getById(instId);
+                entity.setMarca(editMarca);
+                return new ResponseEntity<>(instRepository.save(entity), HttpStatus.CREATED);
+            }
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
